% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_NEON.R
\name{reformat_NEON}
\alias{reformat_NEON}
\title{Clean NEON metabolism parameter data and prep for two-station stream metabolism modeling}
\usage{
reformat_NEON(
  data,
  startdate,
  enddate,
  calculate_light = TRUE,
  single_station_vars
)
}
\arguments{
\item{data}{dataset}

\item{startdate}{start date for filtering}

\item{enddate}{end date for filtering}

\item{calculate_light}{logical, should the function calculate light from geographic position}

\item{single_station_vars}{character string, which variables were measured at only one location}
}
\value{
dataframe of formatted parameter data
}
\description{
Clean NEON metabolism parameter data and prep for two-station stream metabolism modeling
}
